Add Exception

+++ social-network-backend/src/main/java/com/kt/social/auth/service/impl/PermissionServiceImpl.java
@@ -3,6 +3,8 @@ package com.kt.social.auth.service.impl;
 import com.kt.social.auth.model.Permission;
 import com.kt.social.auth.repository.PermissionRepository;
 import com.kt.social.auth.service.PermissionService;
+import com.kt.social.common.exception.BadRequestException;
+import com.kt.social.common.exception.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,7 +19,7 @@ public class PermissionServiceImpl implements PermissionService {
     @Override
     public Permission create(Permission permission) {
         if (permissionRepository.existsByName(permission.getName())) {
-            throw new IllegalArgumentException("Permission already exists: " + permission.getName());
+            throw new BadRequestException("Permission already exists: " + permission.getName());
         }
         return permissionRepository.save(permission);
     }
@@ -30,7 +32,7 @@ public class PermissionServiceImpl implements PermissionService {
     @Override
     public Permission update(Long id, Permission permission) {
         Permission existingPermission = permissionRepository.findById(permission.getId())
-                .orElseThrow(() -> new RuntimeException("Permission not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Permission not found"));
 
         existingPermission.setName(permission.getName());
         existingPermission.setDescription(permission.getDescription());
@@ -40,6 +42,8 @@ public class PermissionServiceImpl implements PermissionService {
 
     @Override
     public void delete(Long id) {
+        if (!permissionRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Không tìm thấy quyền (Permission) với ID: " + id);
+        }
         permissionRepository.deleteById(id);
-    }
-}
+    }}
